+++ apptentive-android-sdk/src/com/apptentive/android/sdk/Apptentive.java
@@ -24,6 +24,7 @@ import com.apptentive.android.sdk.module.messagecenter.ApptentiveMessageCenter;
 import com.apptentive.android.sdk.module.messagecenter.MessageManager;
 import com.apptentive.android.sdk.module.messagecenter.UnreadMessagesListener;
 import com.apptentive.android.sdk.lifecycle.ActivityLifecycleManager;
+import com.apptentive.android.sdk.module.metric.MetricModule;
 import com.apptentive.android.sdk.module.rating.IRatingProvider;
 import com.apptentive.android.sdk.module.survey.OnSurveyFinishedListener;
 import com.apptentive.android.sdk.module.survey.SurveyManager;
@@ -68,30 +69,13 @@ public class Apptentive {
 	 * @param activity The Activity from which this method is called.
 	 */
 	public static void onStart(Activity activity) {
-		init(activity);
-		ActivityLifecycleManager.activityStarted(activity);
-	}
-
-	/**
-	 * Reserved for future use.
-	 * @param activity The Activity from which this method is called.
-	 */
-	public static void onResume(Activity activity) {
-	}
-
-	/**
-	 * Reserved for future use.
-	 * @param activity The Activity from which this method is called.
-	 * @param hasFocus true if the activity is coming into focus, else false.
-	 */
-	public static void onWindowFocusChanged(Activity activity, boolean hasFocus) {
-	}
-
-	/**
-	 * Reserved for future use.
-	 * @param activity The Activity from which this method is called.
-	 */
-	public static void onPause(Activity activity) {
+		try {
+			init(activity);
+			ActivityLifecycleManager.activityStarted(activity);
+		} catch (Exception e) {
+			Log.w("Error starting Apptentive Activity.", e);
+			MetricModule.sendError(activity.getApplicationContext(), e, null, null);
+		}
 	}
 
 	/**
@@ -99,15 +83,13 @@ public class Apptentive {
 	 * @param activity The Activity from which this method is called.
 	 */
 	public static void onStop(Activity activity) {
-		ActivityLifecycleManager.activityStopped(activity);
-		NetworkStateReceiver.clearListeners();
-	}
-
-	/**
-	 * Reserved for future use.
-	 * @param activity The Activity from which this method is called.
-	 */
-	public static void onDestroy(Activity activity) {
+		try {
+			ActivityLifecycleManager.activityStopped(activity);
+			NetworkStateReceiver.clearListeners();
+		} catch (Exception e) {
+			Log.w("Error stopping Apptentive Activity.", e);
+			MetricModule.sendError(activity.getApplicationContext(), e, null, null);
+		}
 	}
 
 
@@ -326,9 +308,15 @@ public class Apptentive {
 	 * Calling this method will display the rating flow's first dialog if the conditions you have specified at
 	 * apptentive.com for this app have been met. Otherwise it will return immediately and have no side effect.
 	 * @param activity The activity from which this set of dialogs is launched.
+	 * @return True if the rating flow was shown, else false.
 	 */
-	public static void showRatingFlowIfConditionsAreMet(Activity activity) {
-		RatingModule.getInstance().run(activity);
+	public static boolean showRatingFlowIfConditionsAreMet(Activity activity) {
+		try {
+			return RatingModule.getInstance().run(activity);
+		} catch (Exception e) {
+			MetricModule.sendError(activity.getApplicationContext(), e, null, null);
+		}
+		return false;
 	}
 
 	// ****************************************************************************************
@@ -342,7 +330,24 @@ public class Apptentive {
 	 * @param activity The Activity from which to launch the Message Center
 	 */
 	public static void showMessageCenter(Activity activity) {
-		showMessageCenter(activity, true);
+		showMessageCenter(activity, true, null);
+	}
+
+	/**
+	 * Opens the Apptentive Message Center UI Activity, and allows custom data to be sent with the next message the user
+	 * sends. If the user sends multiple messages, this data will only be sent with the first message sent after this
+	 * method is invoked. Additional invocations of this method with custom data will repeat this process.
+	 *
+	 * @param activity The Activity from which to launch the Message Center
+	 * @param customData A Map of key/value Strings that will be sent with the next message.
+	 */
+	public static void showMessageCenter(Activity activity, Map<String, String> customData) {
+		try {
+			showMessageCenter(activity, true, customData);
+		} catch (Exception e) {
+			Log.w("Error starting Apptentive Activity.", e);
+			MetricModule.sendError(activity.getApplicationContext(), e, null, null);
+		}
 	}
 
 	/**
@@ -359,7 +364,12 @@ public class Apptentive {
 	 * @return The number of unread messages.
 	 */
 	public static int getUnreadMessageCount(Context context) {
-		return MessageManager.getUnreadMessageCount(context);
+		try {
+			return MessageManager.getUnreadMessageCount(context);
+		} catch (Exception e) {
+			MetricModule.sendError(context.getApplicationContext(), e, null, null);
+		}
+		return 0;
 	}
 
 	// ****************************************************************************************
@@ -374,7 +384,12 @@ public class Apptentive {
 	 * @return True if a survey can be shown, else false.
 	 */
 	public static boolean isSurveyAvailable(Context context, String... tags) {
-		return SurveyManager.isSurveyAvailable(context, tags);
+		try {
+			return SurveyManager.isSurveyAvailable(context, tags);
+		} catch (Exception e) {
+			MetricModule.sendError(context.getApplicationContext(), e, null, null);
+		}
+		return false;
 	}
 
 	/**
@@ -385,7 +400,12 @@ public class Apptentive {
 	 * @return True if a survey was shown, else false.
 	 */
 	public static boolean showSurvey(Activity activity, OnSurveyFinishedListener listener, String... tags) {
-		return SurveyManager.showSurvey(activity, listener, tags);
+		try {
+			return SurveyManager.showSurvey(activity, listener, tags);
+		} catch (Exception e) {
+			MetricModule.sendError(activity.getApplicationContext(), e, null, null);
+		}
+		return false;
 	}
 
 	// ****************************************************************************************
@@ -489,7 +509,7 @@ public class Apptentive {
 			asyncFetchConversationToken(context);
 		} else {
 			asyncFetchAppConfiguration(context);
-			SurveyManager.asynchFetchAndStoreSurveysIfCacheExpired(context);
+			SurveyManager.asyncFetchAndStoreSurveysIfCacheExpired(context);
 		}
 
 		// TODO: Do this on a dedicated thread if it takes too long. Some HTC devices might take like 30 seconds I think.
@@ -527,7 +547,7 @@ public class Apptentive {
 		PayloadSendWorker.start(context);
 	}
 
-	private static void onVersionChanged(Context context, int previousVersion, int currentVersion) {
+	private static void onVersionChanged(Context context, @SuppressWarnings("unused") int previousVersion, @SuppressWarnings("unused") int currentVersion) {
 		RatingModule.getInstance().onAppVersionChanged(context);
 		AppRelease appRelease = AppReleaseManager.storeAppReleaseAndReturnDiff(context);
 		if (appRelease != null) {
@@ -537,12 +557,22 @@ public class Apptentive {
 	}
 
 	private synchronized static void asyncFetchConversationToken(final Context context) {
-		new Thread() {
+		Thread thread = new Thread() {
 			@Override
 			public void run() {
 				fetchConversationToken(context);
 			}
-		}.start();
+		};
+		Thread.UncaughtExceptionHandler handler = new Thread.UncaughtExceptionHandler() {
+			@Override
+			public void uncaughtException(Thread thread, Throwable throwable) {
+				Log.w("Caught UncaughtException in thread \"%s\"", throwable, thread.getName());
+				MetricModule.sendError(context.getApplicationContext(), throwable, null, null);
+			}
+		};
+		thread.setUncaughtExceptionHandler(handler);
+		thread.setName("Apptentive-FetchConversationToken");
+		thread.start();
 	}
 
 	/**
@@ -576,7 +606,7 @@ public class Apptentive {
 				}
 				// Try to fetch app configuration, since it depends on the conversation token.
 				asyncFetchAppConfiguration(context);
-				SurveyManager.asynchFetchAndStoreSurveysIfCacheExpired(context);
+				SurveyManager.asyncFetchAndStoreSurveysIfCacheExpired(context);
 			} catch (JSONException e) {
 				Log.e("Error parsing ConversationToken response json.", e);
 			}
@@ -622,11 +652,21 @@ public class Apptentive {
 	}
 
 	private static void asyncFetchAppConfiguration(final Context context) {
-		new Thread() {
+		Thread thread = new Thread() {
 			public void run() {
 				fetchAppConfiguration(context, GlobalInfo.isAppDebuggable);
 			}
-		}.start();
+		};
+		Thread.UncaughtExceptionHandler handler = new Thread.UncaughtExceptionHandler() {
+			@Override
+			public void uncaughtException(Thread thread, Throwable throwable) {
+				Log.e("Caught UncaughtException in thread \"%s\"", throwable, thread.getName());
+				MetricModule.sendError(context.getApplicationContext(), throwable, null, null);
+			}
+		};
+		thread.setUncaughtExceptionHandler(handler);
+		thread.setName("Apptentive-FetchAppConfiguration");
+		thread.start();
 	}
 
 	/**
@@ -634,9 +674,9 @@ public class Apptentive {
 	 * @param activity The Activity from which to launch the Message Center
 	 * @param forced True if opened manually. False if opened from ratings flow.
 	 */
-	static void showMessageCenter(Activity activity, boolean forced) {
+	static void showMessageCenter(Activity activity, boolean forced, Map<String, String> customData) {
 		MessageManager.createMessageCenterAutoMessage(activity, forced);
-		ApptentiveMessageCenter.show(activity, forced);
+		ApptentiveMessageCenter.show(activity, forced, customData);
 	}
 
 	/**