+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 2013-??-?? skykelsey  next
 --------------------------
 
-New Features:
+Major changes:
+
 * Ability to pass Apptentive custom integration details. For instance, you can pass in the app's Urban Airship App ID, allowing us to send push notifications to your device.
+
+2013-11-13 skykelsey 1.2.3
+--------------------------
+
+Major changes:
+
+* Add method to API. New way of opening Message Center allowing you to include custom data to be sent with the next message the user sends while the message center is open.
+* Implemented an internal error trapping and reporting system. Exceptions in Apptentive code are now caught and logged to our server so we can maintain a reliable app experience and proactively fix hard to reproduce bugs.
+
+Bugs Fixed:
+
+* Fixed a potential crash in survey fetching code.
 
 2013-11-05 skykelsey 1.2.2
 --------------------------